// src/types/api.ts
// let's imagine this file is autogenerated from the backend
// ideally, we want to keep these api related types in sync
// with the backend instead of manually writing them out

export type BaseEntity = {
  id: string;
  createdAt: number;
};

export type Entity<T> = {
  [K in keyof T]: T[K];
} & BaseEntity;

export type Meta = {
  page: number;
  total: number;
  totalPages: number;
};

export type User = Entity<{
  firstName: string;
  lastName: string;
  email: string;
}>;

export type AuthResponse = {
  jwt: string;
  user: User;
};

export type TodoTime = {
  hour: number;
  minutes: number;
};

export type Todo = Entity<{
  title: string;
  description: string;
  tags: string[];
  color: string | null;
  status: 'pending' | 'in-progress' | 'completed' | 'overdue';
  priority: 'low' | 'medium' | 'high';
  time: {
    start: TodoTime;
    end: TodoTime;
  } | null;
  // Only with the precision to the day
  due?: string | null;
  updatedAt: string;
  completedAt?: string | null;
  isRecurring: boolean;
  // e.g "RRULE:FREQ=WEEKLY;BYDAY=MO,WE,FR"
  recurrenceRule?: string | null;
}>;

export type Days = Todo[][];